+++ dart/lib/src/event_processor/enricher/io_enricher_event_processor.dart
@@ -100,6 +100,9 @@ class IoEnricherEventProcessor implements EnricherEventProcessor {
           exception: exception,
           stackTrace: stackTrace,
         );
+        if (_options.automatedTestMode) {
+          rethrow;
+        }
       }
     }