+++ dart/test/load_dart_debug_images_integration_test.dart
@@ -7,6 +7,7 @@ import 'package:test/test.dart';
 
 import 'mocks/mock_platform.dart';
 import 'mocks/mock_platform_checker.dart';
+import 'test_utils.dart';
 
 void main() {
   group(LoadDartDebugImagesIntegration, () {
@@ -92,7 +93,7 @@ isolate_dso_base: 10000000
 }
 
 class Fixture {
-  final options = SentryOptions(dsn: 'https://public@sentry.example.com/1');
+  final options = defaultTestOptions();
 
   Fixture() {
     final integration = LoadDartDebugImagesIntegration();