+++ flutter/ios/Classes/SentryFlutterReplayScreenshotProvider.m
@@ -19,9 +19,12 @@
 - (void)imageWithView:(UIView *_Nonnull)view
               options:(id<SentryRedactOptions> _Nonnull)options
            onComplete:(void (^_Nonnull)(UIImage *_Nonnull))onComplete {
+  // Replay ID may be null if session replay is disabled.
+  // Replay is still captured for on-error replays.
+  NSString *replayId = [PrivateSentrySDKOnly getReplayId];
   [self->channel
       invokeMethod:@"captureReplayScreenshot"
-         arguments:@{@"replayId" : [PrivateSentrySDKOnly getReplayId]}
+         arguments:@{@"replayId" : replayId ? replayId : [NSNull null]}
             result:^(id value) {
               if (value == nil) {
                 NSLog(@"SentryFlutterReplayScreenshotProvider received null "
@@ -33,6 +36,11 @@
                     (FlutterStandardTypedData *)value;
                 UIImage *image = [UIImage imageWithData:typedData.data];
                 onComplete(image);
+              } else if ([value isKindOfClass:[FlutterError class]]) {
+                FlutterError *error = (FlutterError *)value;
+                NSLog(@"SentryFlutterReplayScreenshotProvider received an "
+                      @"error: %@. Cannot capture a replay screenshot.",
+                      error.message);
               } else {
                 NSLog(@"SentryFlutterReplayScreenshotProvider received an "
                       @"unexpected result. "