+++ flutter/lib/src/event_processor/replay_event_processor.dart
@@ -5,9 +5,10 @@ import 'package:sentry/sentry.dart';
 import '../native/sentry_native_binding.dart';
 
 class ReplayEventProcessor implements EventProcessor {
+  final Hub _hub;
   final SentryNativeBinding _binding;
 
-  ReplayEventProcessor(this._binding);
+  ReplayEventProcessor(this._hub, this._binding);
 
   @override
   Future<SentryEvent?> apply(SentryEvent event, Hint hint) async {
@@ -15,7 +16,12 @@ class ReplayEventProcessor implements EventProcessor {
         event.exceptions?.isNotEmpty == true) {
       final isCrash =
           event.exceptions!.any((e) => e.mechanism?.handled == false);
-      await _binding.captureReplay(isCrash);
+      final replayId = await _binding.captureReplay(isCrash);
+      // If session replay is disabled, this is the first time we receive the ID.
+      _hub.configureScope((scope) {
+        // ignore: invalid_use_of_internal_member
+        scope.replayId = replayId;
+      });
     }
     return event;
   }