+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -5,7 +5,6 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
-import '../integrations/integrations.dart';
 import '../native/native_frames.dart';
 import '../native/sentry_native_binding.dart';
 import 'time_to_display_tracker.dart';
@@ -244,13 +243,10 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     String? name = _getRouteName(route);
     final arguments = route?.settings.arguments;
 
-    if (name == null) {
+    if (name == null || (name == '/')) {
       return;
     }
 
-    if (name == '/') {
-      name = rootScreenName;
-    }
     final transactionContext = SentryTransactionContext(
       name,
       SentrySpanOperations.uiLoad,
@@ -331,6 +327,9 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
         exception: exception,
         stackTrace: stacktrace,
       );
+      if (_hub.options.automatedTestMode) {
+        rethrow;
+      }
     } finally {
       await transaction?.finish();
       if (clearAfter) {
@@ -348,15 +347,6 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
 
       bool isAppStart = routeName == '/';
       DateTime startTimestamp = _hub.options.clock();
-      DateTime? endTimestamp;
-
-      if (isAppStart) {
-        final appStartInfo = await NativeAppStartIntegration.getAppStartInfo();
-        if (appStartInfo == null) return;
-
-        startTimestamp = appStartInfo.start;
-        endTimestamp = appStartInfo.end;
-      }
 
       await _startTransaction(route, startTimestamp);
 
@@ -365,12 +355,11 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
         return;
       }
 
-      if (isAppStart && endTimestamp != null) {
-        await _timeToDisplayTracker?.trackAppStartTTD(transaction,
-            startTimestamp: startTimestamp, endTimestamp: endTimestamp);
-      } else {
-        await _timeToDisplayTracker?.trackRegularRouteTTD(transaction,
-            startTimestamp: startTimestamp);
+      if (!isAppStart) {
+        await _timeToDisplayTracker?.trackRegularRouteTTD(
+          transaction,
+          startTimestamp: startTimestamp,
+        );
       }
     } catch (exception, stacktrace) {
       _hub.options.logger(
@@ -379,6 +368,9 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
         exception: exception,
         stackTrace: stacktrace,
       );
+      if (_hub.options.automatedTestMode) {
+        rethrow;
+      }
     } finally {
       _clear();
     }
@@ -391,9 +383,6 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     _completedDisplayTracking = Completer();
     _timeToDisplayTracker?.clear();
   }
-
-  @internal
-  static const String rootScreenName = 'root /';
 
   bool _isRouteIgnored(Route<dynamic> route) {
     return _ignoreRoutes.isNotEmpty &&